[TS migration] Migrate withFullTransactionOrNotFound HOC
